+++ src/main/java/com/ffuntree/ffunfun/data/user/User.java
@@ -1,4 +1,4 @@
-package com.ffuntree.ffunfun.data;
+package com.ffuntree.ffunfun.data.user;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,7 +8,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
