+++ src/main/java/com/ffuntree/ffunfun/data/user/UserSignInDto.java
@@ -1,6 +1,5 @@
-package com.ffuntree.ffunfun.data.dto;
+package com.ffuntree.ffunfun.data.user;
 
-import com.ffuntree.ffunfun.data.User;
 import lombok.Getter;
 
 @Getter
